zero: remove power-HAL-controlled FP power management
As the FP-sensor has a strang behaviour when the PM-interface is being used, with the old version failing
on newer firmwares and the new PM-interface working just on newer firmwares.

It seems that the sleep-pin is the breaking point here, probably being broken on older firmwares, thus
setting it there breaks the sensor like when not setting it on newer bootloaders.

Thus, disable the power-HAL-controlled fingerprint power-management and replace it with some
Android-based solution in the future.

Change-Id: Ibcb75c56b9b0c3c346f73e5ca385ad38a714ed13